Rreefactor inputfields to be rendered in the InputWrapper component that will be responsible to render error messages
