+++ src/Form/types/DatePicker/DatePicker.styled.js
@@ -1,20 +1,31 @@
 import styled from 'styled-components';
-import {
-  KeyboardDatePicker,
-} from '@material-ui/pickers';
+import { KeyboardDatePicker } from '@material-ui/pickers';
 
 export const DatePicker = styled(KeyboardDatePicker)`
     .MuiOutlinedInput-root {
         background: ${({ theme }) => theme.white};
+        padding-right: 0!important;
     }
+
     .MuiInputBase-root input {
         box-sizing: border-box;
         min-height: 4.7rem; 
         font-size: 1.6rem;    
+        
     }
+
+
     fieldset {   
-        border: 0.1rem solid!important;
+        border: 0rem solid!important;
         border-color: ${({ theme }) => `${theme.gray200}!important`};
         border-radius: 5px!important;
+        
+    }
+
+    input {
+        padding: 0!important;
+        margin: 0!important;
+        min-height: 2.4rem!important;
+
     }
 `;
