+++ src/Form/types/Label.js
@@ -1,62 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import * as Tooltip from '../../Tooltip';
-
-import {
-  TooltipIcon,
-  Main,
-  Label as InputLabel,
-  Header,
-  Wrapper,
-} from './Text.styled';
 
 const propTyps = {
   value: PropTypes.string,
-  label: PropTypes.string,
-  name: PropTypes.string.isRequired,
-  tooltip: PropTypes.string.isRequired,
-  noLabel: PropTypes.bool,
 };
 
 const defaultProps = {
   value: '',
-  label: '',
-  noLabel: false,
 };
 
-const Label = (props) => {
-  const {
-    label,
-    name,
-    tooltip,
-    noLabel,
-    value,
-  } = props;
-
-  const [labelValue, setLabelValue] = useState('');
-
-  useEffect(() => {
-    setLabelValue(value);
-  }, [value]);
-
-  return (
-    <Main>
-      { noLabel === false && (
-      <Header>
-        <InputLabel>{label || name}</InputLabel>
-        { tooltip && (
-        <Tooltip.Middle tip={tooltip}>
-          <TooltipIcon />
-        </Tooltip.Middle>
-        )}
-      </Header>
-      )}
-      <Wrapper style={{ opacity: '.4' }}>
-        { labelValue }
-      </Wrapper>
-    </Main>
-  );
-};
+const Label = ({ value }) => <>{value}</>;
 
 Label.defaultProps = defaultProps;
 Label.propTypes = propTyps;
