+++ src/Form/types/RadioGroup/index.js
@@ -1,0 +1,3 @@
+import RadioGroup from './RadioGroup';
+
+export default RadioGroup;
