+++ src/Form/types/Select.js
@@ -1,31 +1,17 @@
 import React, { forwardRef, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import * as Icons from '@material-ui/icons';
-import * as Tooltip from '../../Tooltip';
-
-import {
-  Main,
-  TooltipIcon,
-  Label,
-  Header,
-  SelectWrapper,
-} from './Text.styled';
 
 const propTyps = {
   value: PropTypes.string,
-  label: PropTypes.string,
   name: PropTypes.string.isRequired,
-  tooltip: PropTypes.string.isRequired,
   options: PropTypes.arrayOf(PropTypes.instanceOf(Object)).isRequired,
   props: PropTypes.instanceOf(Object),
-  noLabel: PropTypes.bool,
 };
 
 const defaultProps = {
   value: '',
-  label: '',
   props: {},
-  noLabel: false,
 };
 
 
@@ -47,11 +33,9 @@ const getDefaultSort = (sortKeys) => {
 
 const Select = forwardRef((props, ref) => {
   const {
-    label,
     name,
-    tooltip,
     options,
-    noLabel,
+
   } = props;
 
   const [value, setValue] = useState('');
@@ -66,44 +50,32 @@ const Select = forwardRef((props, ref) => {
   }, []);
 
   return (
-    <Main>
-      { noLabel === false && (
-        <Header>
-          <Label>{label || name}</Label>
-          { tooltip && (
-          <Tooltip.Middle tip={tooltip}>
-            <TooltipIcon />
-          </Tooltip.Middle>
-          )}
-        </Header>
-      )}
-      <SelectWrapper>
-        <select
-          {...props.props}
-          type="select"
-          onChange={({ target }) => setValue(target.value)}
-          value={value}
-          name={name}
-          ref={ref}
-        >
-          { Array.isArray(options) && options
-            .map(({
-              value: optionValue,
-              label: optionLabel,
-              disabled,
-              disabledPreFix,
-              disabledPostFix,
-            }) => (
-              <option key={`${optionLabel}-${value}`} value={optionValue} disabled={disabled}>
-                {disabled && disabledPreFix}
-                {optionLabel}
-                {disabled && disabledPostFix}
-              </option>
-            ))}
-        </select>
-        <Icons.ArrowDropDown className="down-arrow" fontSize="large" />
-      </SelectWrapper>
-    </Main>
+    <>
+      <select
+        {...props.props}
+        type="select"
+        onChange={({ target }) => setValue(target.value)}
+        value={value}
+        name={name}
+        ref={ref}
+      >
+        { Array.isArray(options) && options
+          .map(({
+            value: optionValue,
+            label: optionLabel,
+            disabled,
+            disabledPreFix,
+            disabledPostFix,
+          }) => (
+            <option key={`${optionLabel}-${value}`} value={optionValue} disabled={disabled}>
+              {disabled && disabledPreFix}
+              {optionLabel}
+              {disabled && disabledPostFix}
+            </option>
+          ))}
+      </select>
+      <Icons.ArrowDropDown className="down-arrow" fontSize="large" />
+    </>
   );
 });
